Generation of a cosine signal with noise and its retrieval using the BiLSTM neural network.
In this project, a simple way of constructing a cosine signal and adding noise to it is presented. Noise is added in two ways. One through a value of the signal-to-noise ratio (SNR) and the other by directly assigning a noise value. Additional data on signal strength in Watts and dB were extracted via known unit converters.
Subsequently, a deep learning neural network called Bidirectional Long-Term Short-Term Memory (BiLSTM) was defined. The neural network was trained with the signal without noise as input and with the signal with noise as output. Next, the neural network was asked to predict the signal with the noise, but also the signal without noise. In both cases the results were very accurate. Finally, the network was asked to provide a copy of the signal even more distorted by the noise, without prior training with it. The results were again very accurate, with the information signal being able to be extracted from the distorted signal.